viewmodel update
