+++ app/src/main/java/com/example/dogadoption/navigation/navigation.kt
@@ -11,30 +11,37 @@ import com.example.dogadoption.ui.DogPicturesScreen
 import com.example.dogadoption.ui.DogPreviewScreen
 import com.example.dogadoption.ui.HomeScreen
 import com.example.dogadoption.ui.UserProfile
+import com.example.dogadoption.viewmodels.DogPicsViewModel
 import com.example.dogadoption.viewmodels.DogViewModel
+import com.example.dogadoption.viewmodels.UserViewModel
 
 @Composable
-fun Navigation(navController: NavHostController, dogViewModel: DogViewModel) {
+fun Navigation(
+    navController: NavHostController,
+    dogViewModel: DogViewModel,
+    picsViewModel: DogPicsViewModel,
+    userViewModel: UserViewModel
+) {
     NavHost(navController, startDestination = Screen.HomeScreen.route) {
         composable(Screen.HomeScreen.route) {
-            HomeScreen(navController, dogViewModel)
+            HomeScreen(navController, userViewModel)
         }
         composable(Screen.ListScreen.route) {
             DogListScreen(navController, dogViewModel)
         }
         composable(Screen.PicsScreen.route) {backStackEntry ->
             val breed = backStackEntry.arguments?.getString("breed") ?: ""
-            DogPicturesScreen(navController, dogViewModel, breed)
+            DogPicturesScreen(navController, picsViewModel, breed)
         }
         composable(
             Screen.PrevScreen.route,
             arguments = listOf(navArgument("imageId") { type = NavType.IntType }))
         { backStackEntry ->
             val imageId = backStackEntry.arguments?.getInt("imageId") ?: 0
-            DogPreviewScreen(navController, dogViewModel, imageId)
+            DogPreviewScreen(navController, picsViewModel, imageId)
         }
         composable(Screen.UserScreen.route) {
-            UserProfile(navController, dogViewModel)
+            UserProfile(navController, userViewModel)
         }
     }
 }
