+++ app/src/main/java/com/example/dogadoption/viewmodels/DogViewModel.kt
@@ -26,27 +26,12 @@ class DogViewModel @Inject constructor(
     private val _dogBreeds = MutableLiveData<List<String>>()
     val dogBreeds: LiveData<List<String>> get() = _dogBreeds
 
-    private val _dogImageData = MutableLiveData<List<DogImages>>(emptyList())
-    val dogImageData: LiveData<List<DogImages>> get() = _dogImageData
-
     private var dogBreedsFetched = false
 
     private val _loading = MutableLiveData(false)
-    val loading: LiveData<Boolean> get() = _loading
-
-    private val _favoriteDogs = MutableLiveData<List<DogImages>>(emptyList())
-    val favoriteDogs: LiveData<List<DogImages>> get() = _favoriteDogs
-
-    private val _user = MutableLiveData<User?>()
-    val user: LiveData<User?> get() = _user
-
-    private val _uiState = MutableLiveData(UserState())
-    val uiState: LiveData<UserState> get() = _uiState
 
     init {
         saveDogBreeds()
-        getUser()
-        getFavoriteDogs()
     }
 
     private fun setLoading(loading: Boolean) {
@@ -97,91 +82,6 @@ class DogViewModel @Inject constructor(
             } finally {
                 setLoading(false)
             }
-        }
-    }
-
-    fun saveDogPictures(breed: String) {
-        viewModelScope.launch(dispatcher) {
-            setLoading(true)
-            try {
-                dogRepository.saveDogPictures(breed)
-                Log.d("DOGVIEWMODEL", "Dog pictures saved to database for $breed")
-            } catch (e: Exception) {
-                Log.e("DOGVIEWMODEL", "Failed to save dog pictures to database: ${e.message}")
-            } finally {
-                setLoading(false)
-            }
-        }
-    }
-
-    fun getDogImages(breed: String) {
-        viewModelScope.launch(dispatcher) {
-            try {
-                val imagesFlow = dogRepository.getDogImages(breed)
-                imagesFlow.collect { dogImageUrl ->
-                    _dogImageData.postValue(dogImageUrl)
-                }
-            } catch (e: Exception) {
-                Log.e("DOGVIEWMODEL", "Failed to fetch dog pictures from database: ${e.message}")
-            }
-        }
-    }
-
-    private fun getFavoriteDogs() {
-        viewModelScope.launch(dispatcher) {
-            dogRepository.getFavoriteDogImages().collect { favouriteDogList ->
-                _favoriteDogs.postValue(favouriteDogList)
-            }
-        }
-    }
-
-    fun toggleFavourite(dogImage: DogImages) {
-        viewModelScope.launch(dispatcher) {
-            try {
-                val updatedDogImage = dogImage.copy(isFavourite = !dogImage.isFavourite)
-                dogRepository.toggleFavourite(updatedDogImage)
-                _favoriteDogs.postValue(
-                    _favoriteDogs.value?.map {
-                        if (it.id == updatedDogImage.id) updatedDogImage else it
-                    }
-                )
-                val allImages = (_dogImageData.value ?: emptyList()) + (_favoriteDogs.value ?: emptyList())
-                _dogImageData.postValue(allImages)
-                Log.d("DOGVIEWMODEL", "Toggled favourite status for image ID: ${dogImage.id}")
-            } catch (e: Exception) {
-                Log.e("DOGVIEWMODEL", "Failed to update favorite status: ${e.message}")
-            }
-        }
-    }
-
-    private fun insertUser(name: String) {
-        viewModelScope.launch(dispatcher) {
-            val user = User(userName = name)
-            dogRepository.insertUser(user)
-            _user.postValue(user)
-            _uiState.postValue(_uiState.value?.copy(userName = name, isAddingUser = false))
-        }
-    }
-
-    private fun getUser() {
-        viewModelScope.launch(dispatcher) {
-            val user = dogRepository.getUser()
-            _user.postValue(user)
-            user?.let {
-                _uiState.postValue(_uiState.value?.copy(userName = it.userName))
-            }
-        }
-    }
-
-    fun onEvent(event: UserEvent) {
-        when (event) {
-            is UserEvent.SetUserName -> _uiState.value = _uiState.value?.copy(userName = event.userName)
-            UserEvent.SaveUser -> {
-                val name = _uiState.value?.userName.orEmpty()
-                insertUser(name)
-            }
-            UserEvent.ShowDialog -> _uiState.value = _uiState.value?.copy(isAddingUser = true)
-            UserEvent.HideDialog -> _uiState.value = _uiState.value?.copy(isAddingUser = false)
         }
     }
 }
